loading texts from server, getting surface annotation data
text titles and content are now loaded from the server. done in both font end and back-end, on the front end text data is loaded when hovering over a text title, and when clicked on the annotations will be loaded (not looking forward to re-doing annotation rendering).

optimized sql queries by not getting the connection pool every time and caching connections for 1 minute. common api request results are stored in redis (will be invalidated once updates are made, e.g. annotations added, texts added to a text group).